* examples/work/work.c: (test), (main): misc changes * liboil/conv/conv_3dnow.c: (conv_f32_s16_3dnow), (conv_s32_f32_3dnow): fix asm * liboil/conv/conv_bitstuff.c: disable brokenness * liboil/liboilprofile.h: fix s390 profiling * liboil/simdpack/Makefile.am: * liboil/simdpack/clip_fast.c: (clip_s16_fast), (clip_s16_fast2), (clip_s32_fast): merge clip_s16.c and clip_s32.c * liboil/simdpack/clip_s16.c: remove * liboil/simdpack/clip_s32.c: remove * testsuite/Makefile.am: glib fixes * testsuite/proto3.c: (check_param): fail if problem * testsuite/stride.c: (main): same

+++ liboil/liboilprofile.h
@@ -28,6 +28,8 @@
 #ifndef _LIBOIL_PROFILE_H_
 #define _LIBOIL_PROFILE_H_
 
+#include <stdint.h>
+
 #define OIL_PROFILE_HIST_LENGTH 10
 
 typedef struct _OilProfile OilProfile;
@@ -84,7 +86,7 @@ static inline unsigned long oil_profile_stamp(void)
 
 static inline unsigned long oil_profile_stamp(void)
 {
-	unsigned int ts;
+	uint64_t ts;
 	__asm__ __volatile__ ("STCK %0\n" : "=m" (ts));
 	return ts;
 }
